+++ spec/lib/pact_broker/diagnostic/resources/heartbeat_spec.rb
@@ -6,16 +6,15 @@ module PactBroker
   module Diagnostic
     module Resources
       describe Heartbeat do
-
         include Rack::Test::Methods
 
         let(:app) { PactBroker::Diagnostic::App.new }
 
         describe "GET /diagnostic/status/heartbeat" do
-
+          let(:rack_env) { { "pactbroker.base_url" => "http://pact-broker"} }
           let(:parsed_response_body) { JSON.parse(subject.body) }
 
-          subject { get "/diagnostic/status/heartbeat"; last_response }
+          subject { get("/diagnostic/status/heartbeat", nil, rack_env) }
 
           it "returns a 200" do
             expect(subject.status).to eq 200
@@ -26,7 +25,7 @@ module PactBroker
           end
 
           it "returns a link to itself" do
-            expect(parsed_response_body["_links"]["self"]["href"]).to eq "http://example.org/diagnostic/status/heartbeat"
+            expect(parsed_response_body["_links"]["self"]["href"]).to eq "http://pact-broker/diagnostic/status/heartbeat"
           end
         end
       end